+++ src/Exceptions/InvalidConfigurationException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace BlissJaspis\Midtrans\Exceptions;
+
+use InvalidArgumentException;
+
+class InvalidConfigurationException extends InvalidArgumentException
+{
+    //
+}